refactor: extract app route actions
